+++ packages/dataviews/src/field-types/stories/index.story.tsx
@@ -614,46 +614,10 @@ const FieldTypeStory = ( {
 				return _fields;
 			}
 
-			return _fields.map( ( field ) => {
-				const editConfig = field.Edit;
-				let newEdit: any;
-
-				if ( typeof editConfig === 'string' ) {
-					newEdit = {
-						control: editConfig,
-						disabled: true,
-					};
-				} else if (
-					typeof editConfig === 'object' &&
-					editConfig !== null
-				) {
-					newEdit = {
-						...editConfig,
-						disabled: true,
-					};
-				} else if ( field.type ) {
-					// boolean is the only field type whose default Edit control is not the same as its type.
-					let control: string = field.type;
-					if ( field.type === 'boolean' ) {
-						control = 'checkbox';
-					}
-					if ( field.elements ) {
-						control = 'select';
-					}
-
-					newEdit = {
-						control,
-						disabled: true,
-					};
-				} else {
-					newEdit = null;
-				}
-
-				return {
-					...field,
-					Edit: newEdit,
-				} as Field< DataType >;
-			} );
+			return _fields.map( ( field ) => ( {
+				...field,
+				isDisabled: true,
+			} ) );
 		}, [ disabled, _fields ] );
 
 		if ( Edit !== 'default' ) {