+++ packages/dataviews/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Enhancements
 
-- DataForm: support disabled controls. [#77090](https://github.com/WordPress/gutenberg/pull/77090)
+- DataForm: support `isDisabled` field property. [#77090](https://github.com/WordPress/gutenberg/pull/77090)
 
 ## 14.0.0 (2026-04-01)